fix: moved password to feature module
