+++ feature/editpassword/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    alias(libs.plugins.mifospay.android.feature)
+    alias(libs.plugins.mifospay.android.library.compose)
+}
+
+android {
+    namespace = "org.mifospay.feature.editpassword"
+}
+
+dependencies {
+    implementation(projects.core.data)
+}
