+++ feature/editpassword/src/main/kotlin/org/mifospay/feature/editpassword/EditPasswordScreen.kt
@@ -1,4 +1,4 @@
-package org.mifospay.password.ui
+package org.mifospay.feature.editpassword
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -29,13 +29,10 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.coroutines.launch
-import org.mifos.mobilewallet.mifospay.password.presenter.EditPasswordUiState
-import org.mifos.mobilewallet.mifospay.password.presenter.EditPasswordViewModel
-import org.mifospay.R
 import org.mifospay.core.designsystem.component.MfPasswordTextField
 import org.mifospay.core.designsystem.component.MifosButton
 import org.mifospay.core.designsystem.component.MifosScaffold
-import org.mifospay.theme.MifosTheme
+import org.mifospay.core.designsystem.theme.MifosTheme
 
 @Composable
 fun EditPasswordScreen(
@@ -90,7 +87,7 @@ fun EditPasswordScreen(
             EditPasswordUiState.Loading -> {}
             EditPasswordUiState.Success -> {
                 coroutineScope.launch {
-                    currentSnackbarHostState.showSnackbar(context.getString(R.string.password_changed_successfully))
+                    currentSnackbarHostState.showSnackbar(context.getString(R.string.feature_editpassword_password_changed_successfully))
                 }
             }
 
@@ -99,7 +96,7 @@ fun EditPasswordScreen(
     }
 
     MifosScaffold(
-        topBarTitle = R.string.change_password,
+        topBarTitle = R.string.feature_editpassword_change_password,
         snackbarHost = {
             SnackbarHost(hostState = snackbarHostState)
         },
@@ -115,7 +112,7 @@ fun EditPasswordScreen(
                         .fillMaxWidth()
                         .padding(horizontal = 16.dp),
                     password = currentPassword,
-                    label = stringResource(R.string.current_password),
+                    label = stringResource(R.string.feature_editpassword_current_password),
                     isError = false,
                     isPasswordVisible = isConfirmPasswordVisible,
                     onTogglePasswordVisibility = {
@@ -128,10 +125,10 @@ fun EditPasswordScreen(
                         .fillMaxWidth()
                         .padding(horizontal = 16.dp),
                     password = newPassword,
-                    label = stringResource(id = R.string.new_password),
+                    label = stringResource(id = R.string.feature_editpassword_new_password),
                     isError = newPassword.isNotEmpty() && newPassword.length < 6,
                     errorMessage = if (newPassword.isNotEmpty() && newPassword.length < 6) stringResource(
-                        id = R.string.password_length_error
+                        id = R.string.feature_editpassword_password_length_error
                     ) else null,
                     isPasswordVisible = isNewPasswordVisible,
                     onTogglePasswordVisibility = { isNewPasswordVisible = !isNewPasswordVisible },
@@ -142,10 +139,10 @@ fun EditPasswordScreen(
                         .fillMaxWidth()
                         .padding(horizontal = 16.dp),
                     password = confirmNewPassword,
-                    label = stringResource(id = R.string.confirm_new_password),
+                    label = stringResource(id = R.string.feature_editpassword_confirm_new_password),
                     isError = newPassword != confirmNewPassword && confirmNewPassword.isNotEmpty(),
                     errorMessage = if (newPassword != confirmNewPassword && confirmNewPassword.isNotEmpty()) stringResource(
-                        id = R.string.password_mismatch_error
+                        id = R.string.feature_editpassword_password_mismatch_error
                     ) else null,
                     isPasswordVisible = isConfirmNewPasswordVisible,
                     onTogglePasswordVisibility = {
@@ -166,7 +163,7 @@ fun EditPasswordScreen(
                             .weight(1f)
                             .padding(8.dp),
                         contentPadding = PaddingValues(16.dp),
-                        content = { Text(text = stringResource(id = R.string.cancel)) }
+                        content = { Text(text = stringResource(id = R.string.feature_editpassword_cancel)) }
                     )
                     MifosButton(
                         modifier = Modifier
@@ -176,7 +173,7 @@ fun EditPasswordScreen(
                             onSave.invoke(currentPassword, newPassword, confirmNewPassword)
                         },
                         contentPadding = PaddingValues(16.dp),
-                        content = { Text(text = stringResource(id = R.string.save)) }
+                        content = { Text(text = stringResource(id = R.string.feature_editpassword_save)) }
                     )
                 }
             }
